pyproject.toml
[tool.poetry]
name = "drf-to-do"
version = "0.1.0"
description = "Test task 'to_do' on DRF"
authors = ["vadim-shalapugin <vadimsalapugin@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_to_do"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
django-rest-framework = "^0.1.0"
redis = "^4.3.4"
drf-yasg = "^1.21.4"
celery = "^5.2.7"
mysqlclient = "^2.1.1"
python-dotenv = "^0.21.0"
ipython = "^8.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
